Move go service under backend folder and add frontend folder for react project.
