+++ .gitignore
@@ -1,1 +1,3 @@
 .idea
+
+/frontend/node_modules
