+++ backend/Dockerfile
@@ -7,6 +7,6 @@ COPY . .
 RUN go build -o ./out/ .
 
 FROM alpine:3
-COPY --from=builder ./docs/swagger.yaml ./docs/
+COPY --from=builder docs/swagger.yaml ./docs/
 COPY --from=builder /out/ .
 ENTRYPOINT ["./Open-Stage"]
