+++ frontend/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "learn-react",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/react": "^11.2.6",
+    "@testing-library/user-event": "^12.8.3",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-scripts": "4.0.3",
+    "web-vitals": "^1.1.1"
+  },
+  "scripts": {
+    "start": "../../../../../../Desktop/Programming/learn-react/node_modules/.bin/react-scripts start",
+    "build": "../../../../../../Desktop/Programming/learn-react/node_modules/.bin/react-scripts build",
+    "test": "../../../../../../Desktop/Programming/learn-react/node_modules/.bin/react-scripts test",
+    "eject": "../../../../../../Desktop/Programming/learn-react/node_modules/.bin/react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
+}
