+++ frontend/src/App.jsx
@@ -1,0 +1,8 @@
+import React  from 'react';
+import RoomSelect from './pages/RoomSelect'
+
+const App = () => {
+    return <RoomSelect />
+}
+
+export default App;
