+++ frontend/src/index.css
@@ -1,0 +1,13 @@
+@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600&display=swap');
+
+html {
+    font-family: 'Poppins', sans-serif;
+    color: #eeeeee;
+    overflow-x: hidden;
+}
+
+html body {
+    background-color: #232931;
+    margin: 0 2vw;
+    padding: 0;
+}
